removed all warnings

+++ include/file.h
@@ -8,7 +8,7 @@ namespace file
 	extern unsigned long asm_func;
 	
 	extern std::string inputText; // Text of the input file
-	extern unsigned long inputLineCount; // How many lines does the input file have
+	extern unsigned int inputLineCount; // How many lines does the input file have
 	extern std::string input;  // Path to input file
 	extern std::string output; // Path to output file
 	extern std::vector<std::string> outputVector; // Vector of lines of code (output file)
